Fix error response
The `error` field is not valid for an API Gateway response. As noted in the AWS docs below, when an error response contains invalid fields, it will always return a `502` error code with a body of `{"message": "Internal server error"}`. This PR updates the error response fields to contain valid fields and a more sensible response message.

- https://docs.aws.amazon.com/lambda/latest/dg/services-apigateway.html#services-apigateway-errors